Starting to implement bond-spin rel. as lists
Starting to implement the relationship between bonds and the spins as lists. The plan is for Lattice to inherit classes Spin and Bond, Spin-->Lattice<--Bond, not Bond-->Spin-->Lattice as it is now. The latter double-stores the bonds as each bond only exists inside a site. Hoping to gain efficiency and speed. The code (equilibration and MC-part) still runs on the old implementation. Testing underway, have implemented more (and simpler) types of lattices to verify the implementation.