Changed from column-major to row-major
Affects only n1, n2 and n3 (and for the fcc also xpos, ypos and zpos). This part of the program is not in use yet, so that shouldn't affect anything yet. Row-major ordering means that the first index changes most slowly while the last index varies most rapidly. Not sure this would affect anything, actually.

Also tried to get an output array of N elements (instead of N/2) by calling a complex-to-complex Fourier transform. That didn't work, must not have grasped something about complex numbers in C++. cmath and complex both imported, using namespace std, etc. That part of the code is commented out for now.