Added option for slow cooling of the system
Turn on/off bool slowcool. The results seem to be more stable with slowcool on. There were some instances where we wound up in metastable states without slowcool, and after turning slowcool on, we wound up in the ground state. This could be coincidental, maybe turning slowcool on will lead to metastable states instead of g.s. in some cases, but hopefully not.