Implementing m(q=2pi[0,1,0])
It compiles, but is not tested yet (not sure how I should test it, other than inspecting the plots). The new code will only run if we don't calculate the spin correlation function or the system is one of my extended versions of the fcc (i.e. 'E' or 'Y').